refactor(core): rename OS fhir search and query files
Ref eng-371
